Add Pokemon suggestion feature when inputting a Pokemon
When typing a Pokemon a suggestion will be presented to the user above
the input when the current input is at least four characters and a
unique remaining Pokemon exists that is prefixed by the current input.
The idea behind this feature is that typing in the Pokemon names (some
of which have difficult to remember spelling) slows down the game
experience so some kind of auto-complete or suggestion can improve this.
A minimum input length and unique prefix constraint avoids accidentally
revealing Pokemon (as typing four or more characters that prefix a
Pokemon should mean the user is aware of the specific Pokemon they are
aiming for).