Show suggestion for exact matches
Suggestions only show if the input is four characters or more to avoid
the game becoming too easy to brute force. However there are two Pokemon
currently (Mew and Muk) that have names that are under four characters.
It can be odd for the player to type a complete Pokemon name and not
receive it as a suggestion, so extra handling has been added where a
suggestion will always show if the input is an exact match for a
remaining Pokemon.